onboarding fixes, flickering fixes on quests page
